LRDOCS-12446 Commerce Order Rule implemented new methods, change to BigDecimal

+++ docs/commerce/latest/en/developer-guide/order-management/implementing-a-custom-order-rule/resources/liferay-x9k1.zip/x9k1-impl/src/main/java/com/acme/x9k1/internal/commerce/order/rule/entry/util/X9K1MinimumQuantityUtil.java
@@ -7,13 +7,13 @@ import com.liferay.portal.kernel.util.UnicodePropertiesBuilder;
 
 public class X9K1MinimumQuantityUtil {
 
-	public static int getMinimumQuantity(COREntry corEntry) {
+	public static Double getMinimumQuantity(COREntry corEntry) {
 		UnicodeProperties typeSettingsUnicodeProperties =
 			UnicodePropertiesBuilder.fastLoad(
 				corEntry.getTypeSettings()
 			).build();
 
-		return GetterUtil.getInteger(
+		return GetterUtil.getDouble(
 			typeSettingsUnicodeProperties.getProperty("minimum-quantity"));
 	}
